pyproject.toml
[tool.poetry]
name = "django-ninja-tutorial"
version = "0.1.0"
description = ""
authors = ["Filip Mularski <mularskif@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
django-ninja = "^1.1.0"
django-extensions = "^3.2.3"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
whitenoise = "^6.6.0"
django-colorfield = "^0.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
